Add DROP TABLE support.
Also factored the AST for each SQL command into its own file.
