+++ parser/drop_table.go
@@ -1,0 +1,8 @@
+// nolint: govet
+package parser
+
+type DropTable struct {
+	Table string `"DROP" "TABLE" @( Ident ( "." Ident )* | QuotedIdent )`
+}
+
+func (d *DropTable) node() {}
