+++ parser/insert.go
@@ -1,0 +1,14 @@
+// nolint: govet
+package parser
+
+type InsertOrReplace struct {
+	Replace   bool              `("INSERT" | @"REPLACE")`
+	Into      string            `"INTO" @( Ident ( "." Ident )* | QuotedIdent )`
+	Values    []*InsertTerminal `"VALUES" "(" @@ ")" ( "," "(" @@ ")" )* `
+	Returning *string           `( "RETURNING" @( "NONE" | "ALL_OLD" ) )?`
+}
+
+type InsertTerminal struct {
+	Value
+	Object *JSONObject `| @@`
+}
