+++ parser/parser.go
@@ -2,7 +2,6 @@
 package parser
 
 import (
-	"bytes"
 	"strconv"
 	"strings"
 
@@ -38,9 +37,9 @@ var (
 
 var keywords = []string{
 	"SELECT", "FROM", "WHERE", "LIMIT", "OFFSET", "INSERT", "INTO", "VALUES", "NOT",
-	"BETWEEN", "AND", "OR", "USE", "INDEX", "ASC", "DESC", "CREATE", "TABLE", "HASH", "RANGE", "PROJECTION",
-	"PROVISIONED", "THROUGHPUT", "READ", "WRITE", "GLOBAL", "LOCAL", "INDEX", "SECONDARY",
-	"RETURNING", "NONE", "ALL_OLD", "UPDATED_OLD", "ALL_NEW", "UPDATED_NEW", "DELETE", "CHECK",
+	"BETWEEN", "AND", "OR", "USE", "INDEX", "ASC", "DESC", "DROP", "CREATE", "TABLE", "HASH",
+	"RANGE", "PROJECTION", "PROVISIONED", "THROUGHPUT", "READ", "WRITE", "GLOBAL", "LOCAL", "INDEX",
+	"SECONDARY", "RETURNING", "NONE", "ALL_OLD", "UPDATED_OLD", "ALL_NEW", "UPDATED_NEW", "DELETE", "CHECK",
 }
 
 func keywordsRe() string {
@@ -88,275 +87,8 @@ type Node interface {
 type AST struct {
 	Select      *Select          `(  @@`
 	Insert      *InsertOrReplace ` | @@`
-	CreateTable *CreateTable     ` | @@ ) ";"?`
-}
-
-type CreateTable struct {
-	Table                 string                 `"CREATE" "TABLE" @( Ident ( "." Ident )* | QuotedIdent ) "("`
-	Entries               []*CreateTableEntry    `@@ ("," @@)* ")"`
-	ProvisionedThroughput *ProvisionedThroughput `@@`
-}
-
-func (c *CreateTable) node() {}
-
-type CreateTableEntry struct {
-	GlobalSecondaryIndex *GlobalSecondaryIndex `  @@`
-	LocalSecondaryIndex  *LocalSecondaryIndex  `| @@`
-	Attr                 *TableAttr            `| @@` // Must be last.
-}
-
-func (c *CreateTableEntry) node() {}
-
-type ProvisionedThroughput struct {
-	ReadCapacityUnits  int64 `"PROVISIONED" "THROUGHPUT" "READ" @Number`
-	WriteCapacityUnits int64 `"WRITE" @Number`
-}
-
-func (p *ProvisionedThroughput) node() {}
-
-type GlobalSecondaryIndex struct {
-	Name                  string                 `"GLOBAL" "SECONDARY"? "INDEX" @(Ident | QuotedIdent)`
-	PartitionKey          string                 `"HASH" "(" @(Ident | QuotedIdent) ")"`
-	SortKey               string                 `("RANGE" "(" @(Ident | QuotedIdent) ")")?`
-	Projection            *Projection            `"PROJECTION" @@`
-	ProvisionedThroughput *ProvisionedThroughput `@@`
-}
-
-func (c *GlobalSecondaryIndex) node() {}
-
-type Projection struct {
-	KeysOnly bool     `  @"KEYS_ONLY"`
-	All      bool     `| @"ALL"`
-	Include  []string `| "INCLUDE" (@(Ident | QuotedIdent) ("," (@(Ident | QuotedIdent)))*)`
-}
-
-func (p *Projection) node() {}
-
-type LocalSecondaryIndex struct {
-	Name       string      `"LOCAL" "SECONDARY"? "INDEX" @(Ident | QuotedIdent)`
-	SortKey    string      `"RANGE" "(" @( Ident ( "." Ident )* | QuotedIdent ) ")"`
-	Projection *Projection `"PROJECTION" @@`
-}
-
-func (c *LocalSecondaryIndex) node() {}
-
-type TableAttr struct {
-	Name string `@(Ident | QuotedIdent)`
-	Type string `@Type`
-	Key  string `(@("HASH" | "RANGE") "KEY")?`
-}
-
-func (c *TableAttr) node() {}
-
-// Select based on http://www.h2database.com/html/grammar.html
-type Select struct {
-	Projection *ProjectionExpression `"SELECT" @@`
-	From       string                `"FROM" @( Ident ( "." Ident )* | QuotedIdent )`
-	Index      *string               `( "USE" "INDEX" "(" @Ident ")" )?`
-	Where      *AndExpression        `( "WHERE" @@ )?`
-	Descending *ScanDescending       `( @"ASC" | @"DESC" )?`
-	Limit      *int                  `( "LIMIT" @Number )?`
-}
-
-type InsertOrReplace struct {
-	Replace   bool              `("INSERT" | @"REPLACE")`
-	Into      string            `"INTO" @( Ident ( "." Ident )* | QuotedIdent )`
-	Values    []*InsertTerminal `"VALUES" "(" @@ ")" ( "," "(" @@ ")" )* `
-	Returning *string           `( "RETURNING" @( "NONE" | "ALL_OLD" ) )?`
-}
-
-type InsertTerminal struct {
-	Value
-	Object *JSONObject `| @@`
-}
-
-func (e *Select) node() {}
-
-type ProjectionExpression struct {
-	All     bool                `  ( @"*" | "document" "(" @"*" ")" )`
-	Columns []*ProjectionColumn `| @@ ( "," @@ )*`
-}
-
-func (e *ProjectionExpression) node() {}
-
-func (e ProjectionExpression) String() string {
-	if e.All {
-		return ""
-	}
-	buf := &bytes.Buffer{}
-	buf.WriteString(e.Columns[0].String())
-	for _, p := range e.Columns[1:] {
-		buf.WriteString(", ")
-		buf.WriteString(p.String())
-	}
-	return buf.String()
-}
-
-type ProjectionColumn struct {
-	Function     *FunctionExpression `  @@`
-	DocumentPath *DocumentPath       `| @@`
-}
-
-func (c *ProjectionColumn) node() {}
-
-func (c ProjectionColumn) String() string {
-	if c.DocumentPath != nil {
-		return c.DocumentPath.String()
-	}
-	if c.Function != nil {
-		return c.Function.String()
-	}
-	return ""
-}
-
-type ConditionExpression struct {
-	Or []*AndExpression `@@ ( "OR" @@ )*`
-}
-
-func (e *ConditionExpression) node() {}
-
-type AndExpression struct {
-	And []*Condition `@@ ( "AND" @@ )*`
-}
-
-func (e *AndExpression) node() {}
-
-type Condition struct {
-	Parenthesized *ConditionExpression `  "(" @@ ")"`
-	Not           *NotCondition        `| "NOT" @@`
-	Operand       *ConditionOperand    `| @@`
-	Function      *FunctionExpression  `| @@`
-}
-
-func (e *Condition) node() {}
-
-type NotCondition struct {
-	Condition *Condition `@@`
-}
-
-func (e *NotCondition) node() {}
-
-type FunctionExpression struct {
-	Function string              `@Ident`
-	Args     []*FunctionArgument `"(" @@ ( "," @@ )* ")"`
-}
-
-func (f *FunctionExpression) node() {}
-
-func (f *FunctionExpression) FirstArgIsRef() bool {
-	return len(f.Args) > 0 && f.Args[0].DocumentPath != nil
-}
-
-func (f *FunctionExpression) String() string {
-	buf := &bytes.Buffer{}
-	buf.WriteString(f.Function)
-	buf.WriteRune('(')
-	for i, arg := range f.Args {
-		if i != 0 {
-			buf.WriteString(", ")
-		}
-		buf.WriteString(arg.String())
-	}
-	buf.WriteRune(')')
-	return buf.String()
-}
-
-type FunctionArgument struct {
-	DocumentPath *DocumentPath `  @@`
-	Value        *Value        `| @@`
-}
-
-func (a *FunctionArgument) node() {}
-
-func (a FunctionArgument) String() string {
-	if a.DocumentPath != nil {
-		return a.DocumentPath.String()
-	}
-	if a.Value != nil {
-		return a.Value.String()
-	}
-	return ""
-}
-
-type ConditionOperand struct {
-	Operand      *DocumentPath `@@`
-	ConditionRHS *ConditionRHS `@@`
-}
-
-func (c *ConditionOperand) node() {}
-
-type ConditionRHS struct {
-	Compare *Compare `  @@`
-	Between *Between `| "BETWEEN" @@`
-	In      *In      `| "IN" "(" @@ ")"`
-}
-
-func (c *ConditionRHS) node() {}
-
-type In struct {
-	Values []*Value `@@ ( "," @@ )*`
-}
-
-func (i *In) node() {}
-
-type Compare struct {
-	Operator string   `@( "<>" | "<=" | ">=" | "=" | "<" | ">" | "!=" )`
-	Operand  *Operand `@@`
-}
-
-func (c *Compare) node() {}
-
-type Between struct {
-	Start *Operand `@@`
-	End   *Operand `"AND" @@`
-}
-
-func (b *Between) node() {}
-
-type Operand struct {
-	Value     *Value        `  @@`
-	SymbolRef *DocumentPath `| @@`
-}
-
-func (o *Operand) node() {}
-
-type DocumentPath struct {
-	Fragment []*PathFragment `@@ ( "." @@ )*`
-}
-
-func (p *DocumentPath) node() {}
-
-// String marshals the DocumentPath into a human readable format. Do not use this function when marshaling
-// to expressions, because substitutions need to be applied first for reserved words.
-func (p DocumentPath) String() string {
-	buf := &bytes.Buffer{}
-	buf.WriteString(p.Fragment[0].String())
-	for _, f := range p.Fragment[1:] {
-		buf.WriteRune('.')
-		buf.WriteString(f.String())
-	}
-	return buf.String()
-}
-
-type PathFragment struct {
-	Symbol  string `( @Ident | @QuotedIdent )`
-	Indexes []int  `( "[" @Number "]" )*`
-}
-
-func (p PathFragment) node() {}
-
-func (p PathFragment) String() string {
-	if len(p.Indexes) == 0 {
-		return p.Symbol
-	}
-	buf := &bytes.Buffer{}
-	buf.WriteString(p.Symbol)
-	for _, idx := range p.Indexes {
-		buf.WriteRune('[')
-		buf.WriteString(strconv.Itoa(idx))
-		buf.WriteRune(']')
-	}
-	return buf.String()
+	CreateTable *CreateTable     ` | @@`
+	DropTable   *DropTable       ` | @@ ) ";"?`
 }
 
 type JSONObjectEntry struct {
