+++ parser/testdata/golden/queries.27.golden.go
@@ -1,0 +1,8 @@
+parser.row{
+  Query: "DROP TABLE movies;",
+  AST: &parser.AST{
+    DropTable: &parser.DropTable{
+      Table: "movies",
+    },
+  },
+}
