+++ parser/testdata/queries.sql
@@ -40,3 +40,5 @@ CREATE TABLE movies (title STRING, year NUMBER) PROVISIONED THROUGHPUT READ 1 WR
 CREATE TABLE movies (title STRING HASH KEY, year NUMBER RANGE KEY) PROVISIONED THROUGHPUT READ 1 WRITE 1;
 CREATE TABLE movies (title STRING, year NUMBER, GLOBAL SECONDARY INDEX year_title HASH(year) RANGE(title) PROJECTION ALL PROVISIONED THROUGHPUT READ 1 WRITE 1) PROVISIONED THROUGHPUT READ 1 WRITE 1;
 CREATE TABLE movies (title STRING, year NUMBER, LOCAL SECONDARY INDEX year_index RANGE(year) PROJECTION ALL) PROVISIONED THROUGHPUT READ 1 WRITE 1;
+-- drop table
+DROP TABLE movies;
