+++ querybuilder/create_table.go
@@ -1,3 +1,4 @@
+// nolint: govet
 package querybuilder
 
 import (
